test: wave 4 phase B (2/2) — Account, User, PortalSettings controllers
Adds smoke + submit coverage for the remaining lukewarm controllers:

- AccountController: +7 tests (changePassword, privacy,
  notifications-requires-portal-mod, newsletter, additional, delete,
  merge accounts page renders). Existing 6 tests kept.
- UserController: +8 tests, replacing the misnamed UserControllerCest
  file. New tests target real routes (list, gridView, feed/grid with
  start+sort, detail, initials, send form, guestimage); the deleted
  Cest file targeted /user/create + /user/feed which don't exist on
  the controller.
- PortalSettingsController: +17 tests across the major sub-pages
  (general, appearance, support, portalhome, roomcreation,
  roomcategories, auth/{local,ldap,shib,guest,workspacemembership},
  mailtexts, notifications, privacy, inactive, announcements,
  csvimport). Existing 6 tests kept.

Production fixes surfaced while wiring these tests:

- UserVoter::voteOnAttribute: handle Portal-entity subjects (from
  MapEntity-resolved `subject: 'portal'` IsGranted attributes) without
  forcing intval on the object — Account/notifications and other
  PORTAL_MODERATOR-gated routes used to emit a "could not be converted
  to int" warning here.
- Account/DeleteType: switch IdenticalTo to named-arg form (Symfony 7.3
  deprecation).
- Portal/ServerGeneralType + Portal/SupportType: pass requireTld:false
  to Url constraint (Symfony 7.1 deprecation re: default change to
  true).
- AvatarService::imagettftext call: cast x/y to int explicitly (PHP
  implicit-float-to-int deprecation).

Suite: 474 tests, 2884 assertions, 0 skipped, 0 failures (verified
twice for stability).

+++ src/Services/AvatarService.php
@@ -115,9 +115,9 @@ class AvatarService
         $textWidth = $textBox[2] - $textBox[0];
         $textHeight = $textBox[7] - $textBox[1];
 
-        // Calculate coordinates of the text
-        $x = ($this->imageWidth / 2) - ($textWidth / 2);
-        $y = ($this->imageHeight / 2) - ($textHeight / 2);
+        // Calculate coordinates of the text (imagettftext expects ints)
+        $x = (int) (($this->imageWidth / 2) - ($textWidth / 2));
+        $y = (int) (($this->imageHeight / 2) - ($textHeight / 2));
 
         imagettftext($image, $fontSize, $angle, $x, $y, $colors['text'], $font, $initialString);
